+++ spec/custom_view_spec.rb
@@ -164,7 +164,7 @@ describe 'view' do
   describe "inherited views" do
     it "should support parent views for objects of the subclass" do
       CouchPotato.database.save_document CustomBuild.new(:state => 'success', :time => '2008-01-01')
-      CouchPotato.database.view(CustomBuild.timeline).should have(1).item
+      CouchPotato.database.view(CustomBuild.timeline).size.should == 1
       CouchPotato.database.view(CustomBuild.timeline).first.should be_kind_of(CustomBuild)
     end
   end